Add implementation of [Lockwood and Sei '20].
This commit adds a base configuration for the variational
deep Q-learning algorithm proposed in [Lockwood and Sei '20].
The underlying variational quantum circuit is implement in
PennyLane; this might change in future commits due to
flexibility issues with the framework.

Additionally, this commit fixes a minor bug in the
existing DQN implementation which would lead to a crash
if the sampled batch contained only terminal states.